host-containers: add safeguards against lingering host containers
Now that host-ctr has the ability to rebind to existing host containers.
We want to ensure whenever we enable host containers the container will
be running with its latest configuration.

We utilize `host-ctr`'s clean-up command to clean up any potential
lingering host-container when we're enabling a previously disabled
host-container and whenever a host-container is disabled.